Clap Trap (#732)
* Clap Trap (WIP)

Adds:
Unanimated upscaled rig
"Clap" block
"Change Background Color" block
"Change Hand Colors" block
Input detection
Upscaled sprites
and more :)

* Doll Animation Start

Started animation on the doll's head.

oh yea also everything was misaligned oops will realign eventually

* Doll Head Animations

Completed all the animations for the clap trap doll, added a block to play certain animations (e.g. talking), and started animations on clapping

* Right Arm Fully Animated & Usable

All animations for the right arm are now complete and play properly in the game.

* All Arm Animations Complete

All the animations for the arms (and clap effect) are now done.

* Spotlight

A spotlight was added for the clap cue, can be recolored and works with stacked cues. Still kind of WIP but is fully functional.

* Incoming Objects Start

Sword go woosh

* meow

* Most clap trap objects

Mostly just making this so that I can remove this from my computer to save 5 tasty gigabytes. Not abandoning this 100% but it will be a bit

* Modernize Code

Gotta update all the code that broke in the 9 months of this being dead

* Animation Touch-Ups & Shadows

Touched up all the old animations to be more accurate to the originals (inaccurate framerate, missing frames, misplaced frames). Also added all the shadows that show up on the doll when the spotlight comes on.

* More Animation work and Modernized BG Event

Added animations to move the whole doll on a just/missed input. Fixed bugs with the default hand object not properly playing its animations due to its names previously being changed. Modernized the BG Colors event to have the whole easing dropdown in place of the "Fade?" toggle. Combined the spotlight recoloring section of the BG Colors event with the Hand Colors event to make the Object Colors event.

* New SFX rips

Kievit approved

* New Icons

RIP old icon

* Removed \n<color=#eb5454>[WIP]</color>

Removed \n<color=#eb5454>[WIP]</color>

---------

Co-authored-by: Roshibomb <58871286+Roshibomb@users.noreply.github.com>
Co-authored-by: AstrlJelly <bdlawson115@gmail.com>
Co-authored-by: minenice55 <star.elementa@gmail.com>